Move the milestone constants to scorecard_constant.js
