+++ app/components/ScorecardProgress/ScorecardProgressButtons.js
@@ -13,7 +13,7 @@ import { FontFamily } from '../../assets/stylesheets/theme/font';
 import scorecardHelper from '../../helpers/scorecard_helper';
 import scorecardTracingStepsService from '../../services/scorecard_tracing_steps_service';
 import Scorecard from '../../models/Scorecard';
-import { FINISHED } from '../../constants/milestone_constant';
+import { FINISHED } from '../../constants/scorecard_constant';
 import { getDeviceStyle, containerPadding } from '../../utils/responsive_util';
 import { getMobileFontSizeByPixelRatio } from '../../utils/font_size_util';
