--- app/constants/milestone_constant.js
@@ -1,7 +1,0 @@
-export const DOWNLOADED = 'downloaded';
-export const RUNNING = 'running';
-export const FINISHED = 'finished';
-export const SUBMITTED = 'submitted';
-export const RENEWED = 'renewed';
-export const IN_REVIEW = 'in_review';
-export const COMPLETED = 'completed';
