+++ app/services/scorecardService.js
@@ -10,7 +10,7 @@ import ScorecardReference from '../models/ScorecardReference';
 
 import { scorecardAttributes } from '../utils/scorecard_attributes_util';
 import { handleApiResponse, sendRequestToApi } from './api_service';
-import { IN_REVIEW } from '../constants/milestone_constant';
+import { IN_REVIEW } from '../constants/scorecard_constant';
 
 class ScorecardService {
